encode elrond transactions
