+++ src/families/elrond/logic.ts
@@ -70,20 +70,6 @@ export const getNonce = (a: Account): number => {
   return nonce;
 };
 
-export const encodeESDTTransfer = (t: Transaction, ta: SubAccount): string => {
-  const tokenIdentifierHex = ta.id.split("/")[2];
-  let amountHex = t.amount.toString(16);
-
-  //hex amount length must be even so protocol would treat it as an ESDT transfer
-  if (amountHex.length % 2 !== 0) {
-    amountHex = "0" + amountHex;
-  }
-
-  return Buffer.from(
-    `ESDTTransfer@${tokenIdentifierHex}@${amountHex}`
-  ).toString("base64");
-};
-
 export const computeTransactionValue = async (
   t: Transaction,
   a: Account,
